Split virus_scan.py in multiple files
