+++ iaso/api/form_attachments.py
@@ -9,7 +9,7 @@ from rest_framework.response import Response
 from hat.menupermissions import models as permission
 from iaso.models import Form, FormAttachment, Project
 from iaso.utils.encryption import calculate_md5
-from iaso.utils.virus_scan import scan_uploaded_file_for_virus
+from iaso.utils.virus_scan.clamav import scan_uploaded_file_for_virus
 
 from .common import ModelViewSet, TimestampField
 from .forms import HasFormPermission
