+++ iaso/tests/utils/test_clamav.py
@@ -10,7 +10,7 @@ from django.test import override_settings
 
 from hat import settings
 from iaso.test import TestCase
-from iaso.utils.virus_scan import VirusScanStatus, scan_disk_file_for_virus, scan_uploaded_file_for_virus
+from iaso.utils.virus_scan.clamav import VirusScanStatus, scan_disk_file_for_virus, scan_uploaded_file_for_virus
 
 
 @override_settings(
