+++ plugins/polio/models/base.py
@@ -41,7 +41,7 @@ from iaso.utils.models.soft_deletable import (
     DefaultSoftDeletableManager,
     SoftDeletableModel,
 )
-from iaso.utils.virus_scan import ModelWithFile
+from iaso.utils.virus_scan.model import ModelWithFile
 from plugins.polio.preparedness.parser import open_sheet_by_url
 from plugins.polio.preparedness.spread_cache import CachedSpread
