+++ plugins/polio/tests/api/vaccine_stock_management/test_outgoing_stock_movement_virus_scan.py
@@ -9,7 +9,7 @@ from django.test import override_settings
 from rest_framework import status
 
 from hat import settings
-from iaso.utils.virus_scan import VirusScanStatus
+from iaso.utils.virus_scan.model import VirusScanStatus
 from plugins.polio import models as pm
 from plugins.polio.tests.test_vaccine_stock_management import VaccineStockManagementAPITestCase
